Route mobile push-to-talk through native Rust voice control
Move Android foreground PTT control across the Tauri boundary and preserve Rust-owned turn completion through native playback teardown. Repair the iOS FFI boundary and first-run microphone permission path while keeping iOS background voice withheld.

Constraint: Android native voice must not fall back to WebView microphone capture once the native bridge is packaged.\nRejected: Reusing the desktop voice port | Android has service-owned lifecycle and no generation-bearing UI contract.\nConfidence: medium\nScope-risk: moderate\nDirective: Keep Android device/emulator and iOS Xcode evidence gates separate from compile-time claims.\nTested: UI and Tauri typechecks; focused UI/Tauri Vitest suites; Rust Android contract test; iOS source policy scripts; x86_64 Android debug APK with Temurin 17.\nNot-tested: Android emulator/device runtime, Gateway turn, physical ARM64 acoustic proof, Swift/Xcode final link.

+++ apps/aurora-tauri/src-tauri/ios/AuroraNativePlugin/Sources/AuroraNativePlugin/AuroraIOSVoiceCapture.swift
@@ -35,7 +35,7 @@ public final class AuroraIOSVoiceCapture {
   ) {
     self.session = session
     self.maxChunkSamples = AVAudioFrameCount(max(1, maxChunkSamples))
-    self.state = aurora_ios_audio_state_new(capacityChunks, maxChunkSamples)
+    self.state = aurora_ios_audio_state_new(UInt(capacityChunks), UInt(maxChunkSamples))
   }
 
   deinit {
@@ -69,7 +69,7 @@ public final class AuroraIOSVoiceCapture {
       let result = aurora_ios_audio_state_push_pcm_f32(
         state,
         channels[0],
-        frameCount,
+        UInt(frameCount),
         self.sequence,
         UInt32(format.sampleRate.rounded())
       )
